bugfix: serialise commands
there was a chance that the state update from the device would take to much time and that a new command would already get scheduled and send outdated data back to the device. this now ensures, that every command requires first a state update from the device, unless this feature is explicitly disabled.